pyproject.toml
# SETUP CONFIGURATION.                                                        #
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdib"
authors = [
    {name = "Yang Zhang", email = "yangzhang@u.nus.edu"},
    {name = "Jin Er", email = "er.jin@lfb.rwth-aachen.de"}
]
version = "0.1.0"
description = "SD project"
requires-python = ">=3.10"

####### REQUIRED DEPENDENCIES #######
dependencies = [
    "omegaconf>=2.1.1",
    "diffusers>=0.31.0",
    "transformers>=4.44.2",
    "accelerate>=0.33.0",
    "numpy>=1.23.5,<2.0",
    "torchvision>=0.19.1",
    "open-clip-torch>=2.23.0,<2.26.1",
    "webdataset==0.2.100",
    "gdown==5.2.0",
    "wandb>=0.12.17,<=0.15.9",
    "datasets>=3.1.0",
    "peft>=0.13.2"
]

[tool.setuptools.packages.find]
where = ["src"]  # list of folders that contain the packages (["."] by default)
include = ["sdib"]  # package names should match these glob patterns (["*"] by default)


[project.optional-dependencies]
core = [
    "einops>=0.3.2",
    "imgaug==0.4.0",
    "matplotlib>=3.4.3",
    "opencv-python>=4.5.3.56",
    "pandas>=1.1.0",
    "timm<=1.0.7,>=1.0.7",
    "lightning>=2.2",
    "torch>=2",
    "torchmetrics>=1.3.2",
    "jupyterlab>=3.1.0",
    "torch-fidelity==0.3.0"
]
loggers = [
    "tensorboard",
]
test = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "pytest-mock",
    "pytest-sugar",
]
notebook = [
    "jupyterlab",
    "notebook",
    "jupytext"
]

# Exclude a variety of commonly ignored directories.
# RUFF CONFIGURATION
[tool.ruff]
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv"
]

[tool.black]
line-length = 120
target-version = ['py37', 'py39', 'py310', 'py311']
